renaming the FibonacciPreviews.kt file to FibonacciLightPreviews.kt and its refactoring
